+++ src/app/app-routing.module.ts
@@ -9,13 +9,13 @@ import { OrganizationLayoutComponent } from './layouts/organization-layout.compo
 import { UserLayoutComponent } from './layouts/user-layout.component';
 
 import { AcceptOrganizationComponent } from './accounts/accept-organization.component';
+import { ChangePasswordComponent } from './accounts/change-password.component';
 import { HintComponent } from './accounts/hint.component';
 import { LockComponent } from './accounts/lock.component';
 import { LoginComponent } from './accounts/login.component';
 import { RecoverDeleteComponent } from './accounts/recover-delete.component';
 import { RecoverTwoFactorComponent } from './accounts/recover-two-factor.component';
 import { RegisterComponent } from './accounts/register.component';
-import { ResetMasterPasswordComponent } from './accounts/reset-master-password.component';
 import { SsoComponent } from './accounts/sso.component';
 import { TwoFactorComponent } from './accounts/two-factor.component';
 import { VerifyEmailTokenComponent } from './accounts/verify-email-token.component';
@@ -107,9 +107,9 @@ const routes: Routes = [
                 data: { titleId: 'createAccount' }, // TODO
             },
             {
-                path: 'reset-master-password', component: ResetMasterPasswordComponent,
+                path: 'change-password', component: ChangePasswordComponent,
                 canActivate: [UnauthGuardService],
-                data: { titleId: 'resetMasterPassword' },
+                data: { titleId: 'setMasterPassword' },
             },
             {
                 path: 'hint', component: HintComponent,
