+++ src/locales/en/messages.json
@@ -3164,7 +3164,10 @@
   "taxInfoUpdated": {
     "message": "Tax information updated."
   },
-  "resetMasterPassword": {
-    "message": "Reset Master Password"
+  "setMasterPassword": {
+    "message": "Set Master Password"
+  },
+  "ssoCompleteRegistration": {
+    "message": "In order to complete logging in with SSO, please set a master password below. Make sure to choose a strong password/passphrase and comply with all applied organizational policies."
   }
 }
